Extract TrelloCard into trello_card.dart
