+++ lib/src/sdk/cards/cards.dart
@@ -3,3 +3,4 @@ export 'card.dart';
 export 'card_client.dart';
 export 'card_id.dart';
 export 'card_models.dart';
+export 'trello_card.dart';
